requirements.txt
fastapi
uvicorn[standard]
pillow
requests
